Update block_course_menu.php
Unfortunately the navigation style is theme dependent, so no matter how I change the style in the css of the course menu, the theme overrides it and the white-space:nowrap, will prevent lines from wrapping should the theme decide so. To enforce wrapping consistently over themes, the change needs to be added into the course menu generation file. I narrowed it down to this one change.